+++ src/main/java/subway/models/Station.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.models;
 
 import javax.persistence.*;
