+++ src/main/java/subway/repositories/StationRepository.java
@@ -1,6 +1,7 @@
-package subway;
+package subway.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import subway.models.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
 }
