+++ src/main/java/subway/services/StationService.java
@@ -1,10 +1,14 @@
-package subway;
+package subway.services;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import subway.dto.request.StationRequest;
+import subway.dto.response.StationResponse;
+import subway.models.Station;
+import subway.repositories.StationRepository;
 
 @Service
 @Transactional(readOnly = true)
